Respect neighbor assignment for composite glyphs
I found out that an explicit neighbor assignment for ldot would not work – I still saw the neighbors assigned for l.
The reason for this was RamsaySt always decomposing composite glyphs (via `getBaseGlyph`) – which isn’t a problem for non-composite glyphs, but will lead to a glyph list in which not all pairs work.

This update checks if the glyph name exists in the RamsaySt data first, and only then falls back to the composite’s base glyph.